+++ packages/adp-tooling/test/unit/prompts/change-data-source/index.test.ts
@@ -2,7 +2,7 @@ import { jest } from '@jest/globals';
 import type { ManifestNamespace } from '@sap-ux/project-access';
 import { initI18nProjectValidators } from '@sap-ux/project-input-validator';
 
-const mockFilterDataSourcesByType = jest.fn().mockImplementation((dataSources: Record<string, any>, type: string) => {
+const mockFilterDataSourcesByType = jest.fn().mockImplementation((dataSources, type) => {
     if (!dataSources) {
         return {};
     }
@@ -14,8 +14,8 @@ jest.unstable_mockModule('@sap-ux/project-access', () => ({
     filterDataSourcesByType: mockFilterDataSourcesByType
 }));
 
-const { getPrompts } = await import('../../../../src/prompts/change-data-source/index');
-const i18n = await import('../../../../src/i18n');
+const { getPrompts } = await import('../../../../src/prompts/change-data-source/index.js');
+const i18n = await import('../../../../src/i18n.js');
 
 describe('getPrompts', () => {
     const dataSources = {
@@ -43,7 +43,7 @@ describe('getPrompts', () => {
 
     beforeEach(() => {
         jest.clearAllMocks();
-        mockFilterDataSourcesByType.mockImplementation((ds: Record<string, any>, type: string) => {
+        mockFilterDataSourcesByType.mockImplementation((ds, type) => {
             if (!ds) {
                 return {};
             }