+++ packages/annotation-generator/test/unit/cds-sap-annotations.test.ts
@@ -5,9 +5,9 @@ import { create as createEditor } from 'mem-fs-editor';
 import type { Editor } from 'mem-fs-editor';
 
 import { getProject } from '@sap-ux/project-access';
-import type { AnnotationServiceParameters, GenerateAnnotationsOptions } from '../../src';
-import { generateAnnotations } from '../../src';
-import { adaptFilePath } from '../../src/utils';
+import type { AnnotationServiceParameters, GenerateAnnotationsOptions } from '../../src/index.js';
+import { generateAnnotations } from '../../src/index.js';
+import { adaptFilePath } from '../../src/utils.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));