+++ packages/create/test/unit/cli/get/system.test.ts
@@ -3,8 +3,8 @@ import { Command } from 'commander';
 import type { ToolsLogger } from '@sap-ux/logger';
 import { BackendSystem, SystemType, ConnectionType, AuthenticationType } from '@sap-ux/store';
 
-const mockGetLogger = jest.fn();
-const mockSetLogLevelVerbose = jest.fn();
+const mockGetLogger = jest.fn() as jest.Mock;
+const mockSetLogLevelVerbose = jest.fn() as jest.Mock;
 jest.unstable_mockModule('../../../../src/tracing/logger', () => ({
     getLogger: mockGetLogger,
     setLogLevelVerbose: mockSetLogLevelVerbose