+++ packages/fe-fpm-writer/test/unit/common.test.ts
@@ -3,8 +3,8 @@ import type { Editor } from 'mem-fs-editor';
 import { create as createStorage } from 'mem-fs';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { addExtensionTypes } from '../../src/common/utils';
-import { detectTabSpacing, getJsonSpace } from '../../src/common/file';
+import { addExtensionTypes } from '../../src/common/utils.js';
+import { detectTabSpacing, getJsonSpace } from '../../src/common/file.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));