+++ packages/logger/test/unit/winston-logger/logger.test.ts
@@ -1,16 +1,16 @@
 import 'jest-extended';
-import type { Transport } from '../../../src';
-import { LogLevel, ToolsLogger } from '../../../src';
-import type { ArrayTransportLogEntry } from '../../../src/transports';
+import type { Transport } from '../../../src/index.js';
+import { LogLevel, ToolsLogger } from '../../../src/index.js';
+import type { ArrayTransportLogEntry } from '../../../src/transports/index.js';
 import {
     ConsoleTransport,
     ArrayTransport,
     NullTransport,
     UI5ToolingTransport,
     VSCodeTransport
-} from '../../../src/transports';
-import { NullTransport as WinstonNullTransport } from '../../../src/winston-logger/null-transport';
-import { VSCodeTransport as WinstonVSCodeTransport } from '../../../src/winston-logger/vscode-output-channel-transport';
+} from '../../../src/transports/index.js';
+import { NullTransport as WinstonNullTransport } from '../../../src/winston-logger/null-transport.js';
+import { VSCodeTransport as WinstonVSCodeTransport } from '../../../src/winston-logger/vscode-output-channel-transport.js';
 import winston from 'winston';
 
 jest.mock(