+++ packages/preview-middleware-client/test/unit/adp/sync-views-utils.test.ts
@@ -5,11 +5,11 @@ import {
     isSyncView,
     getSyncViewIds,
     resetSyncViews
-} from '../../../src/adp/sync-views-utils';
+} from '../../../src/adp/sync-views-utils.js';
 import Element from 'mock/sap/ui/core/Element';
 import Log from 'mock/sap/base/Log';
 import ElementRegistry from 'mock/sap/ui/core/ElementRegistry';
-import { CommunicationService } from '../../../src/cpe/communication-service';
+import { CommunicationService } from '../../../src/cpe/communication-service.js';
 import { MessageBarType, showInfoCenterMessage } from '@sap-ux-private/control-property-editor-common';
 
 describe('sync-views-utils', () => {