+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @sap-ux/ui5-application-inquirer
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [fff7490]
+    - @sap-ux/project-access@2.1.1
+    - @sap-ux/inquirer-common@1.0.6
+    - @sap-ux/project-input-validator@1.0.5
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [b326a9a]
+    - @sap-ux/project-access@2.1.0
+    - @sap-ux/inquirer-common@1.0.5
+    - @sap-ux/project-input-validator@1.0.4
+
+## 1.0.4
+
+### Patch Changes
+
+- 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
+- Updated dependencies [21a3de7]
+    - @sap-ux/project-input-validator@1.0.3
+    - @sap-ux/inquirer-common@1.0.4
+    - @sap-ux/project-access@2.0.3
+    - @sap-ux/ui5-info@1.0.1
+
+## 1.0.3
+
+### Patch Changes
+
+- @sap-ux/project-access@2.0.2
+- @sap-ux/inquirer-common@1.0.3
+- @sap-ux/project-input-validator@1.0.2
+
+## 1.0.2
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes