+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 1.0.6
+
+### Patch Changes
+
+- @sap-ux/fiori-generator-shared@1.0.6
+- @sap-ux/ui5-library-writer@1.0.5
+- @sap-ux/ui5-library-inquirer@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- @sap-ux/fiori-generator-shared@1.0.5
+- @sap-ux/ui5-library-writer@1.0.4
+- @sap-ux/ui5-library-inquirer@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
+- Updated dependencies [21a3de7]
+    - @sap-ux/fiori-generator-shared@1.0.4
+    - @sap-ux/fiori-tools-settings@1.0.1
+    - @sap-ux/ui5-library-inquirer@1.0.4
+    - @sap-ux/ui5-library-writer@1.0.3
+    - @sap-ux/nodejs-utils@1.0.2
+    - @sap-ux/ui5-info@1.0.1
+
+## 1.0.3
+
+### Patch Changes
+
+- @sap-ux/ui5-library-writer@1.0.2
+- @sap-ux/fiori-generator-shared@1.0.3
+- @sap-ux/ui5-library-inquirer@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [41f327a]
+    - @sap-ux/nodejs-utils@1.0.1
+    - @sap-ux/fiori-generator-shared@1.0.2
+    - @sap-ux/ui5-library-inquirer@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes